requirements.txt
pytorch
scikit-learn
numpy
pandas
scipy
statistics
tqdm